chore: move t6 xmodel json to proper folder
