+++ src/ObjCommon/Game/T6/XModel/JsonXModel.h
@@ -1,11 +1,7 @@
 #pragma once
 
-#include "Game/T6/T6.h"
-
 #include "Json/JsonCommon.h"
-#include "Json/JsonExtension.h"
 #include <memory>
-#include <nlohmann/json.hpp>
 #include <optional>
 #include <string>
 #include <vector>
